requirements.txt
streamlit==1.8.1
torch==1.12.0
torchvision==0.13.0
Pillow==7.1.2
matplotlib==3.2.2
